experiments/2025-07-05_hybrid_training/requirements.txt
# Requirements for 2025-07-05 Hybrid Training Experiment
# Cross-platform compatible for Mac and Windows

# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.0

# Data processing
numpy>=1.21.0
pandas>=1.5.0
opencv-python>=4.7.0
Pillow>=9.0.0
scikit-learn>=1.2.0

# Configuration and file handling
PyYAML>=6.0
pathlib2>=2.3.7; python_version<"3.4"

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
tqdm>=4.64.0

# Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Logging and experiment tracking
tensorboard>=2.12.0
wandb>=0.14.0

# Development tools
jupyterlab>=3.6.0
ipywidgets>=8.0.0

# Platform-specific optimizations (optional)
# For Apple Silicon Mac users:
# torch-audio>=0.13.0; sys_platform=="darwin" and platform_machine=="arm64"

# For CUDA users (Windows/Linux):
# torch-audio>=0.13.0; sys_platform=="win32" or sys_platform=="linux"

# Additional utilities
psutil>=5.9.0  # For system monitoring
colorama>=0.4.6  # For cross-platform colored terminal output 